Migrate to Swift 4.0.2
