+++ .travis.yml
@@ -16,19 +16,11 @@ matrix:
     - os: linux
       dist: trusty
       sudo: required
-      env: SWIFT_SNAPSHOT=4.0
-    - os: linux
-      dist: trusty
-      sudo: required
       services: docker
-      env: DOCKER_IMAGE=ubuntu:16.04 SWIFT_SNAPSHOT=4.0
+      env: DOCKER_IMAGE=ubuntu:16.04
     - os: osx
-      osx_image: xcode8.3
+      osx_image: xcode9.1
       sudo: required
-    - os: osx
-      osx_image: xcode9
-      sudo: required
-      env: SWIFT_SNAPSHOT=4.0
 
 script:
   - ./build.sh
