--- Package.pins
@@ -1,5 +1,0 @@
-{
-  "autoPin": false,
-  "pins": [],
-  "version": 1
-}
